pyproject.toml
[tool.poetry]
name = "Yumemi"
version = "0.1.0"
description = "Yumemi"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.1.0"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"

[nonebot.plugins]
plugins = ["nonebot_plugin_docs", "nonebot_plugin_test"]
plugin_dirs = ["src/plugins/modules"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
